Changed routing of student pages (check tele). Added responsiveness. TODO: TopNavBar Home button href not working
