+++ apps/web/src/app/dashboard/student/quiz-reports/[slug]/_component/ResultTable.tsx
